+++ CHANGELOG.md
@@ -28,6 +28,8 @@
   - `bedtools/bedtools_sort`: Sorts a feature file (bed/gff/vcf) by chromosome and other criteria (PR #98).
 
 
+* `fastqc`: High throughput sequence quality control analysis tool (PR #92).
+
 ## MINOR CHANGES
 
 * `busco` components: update BUSCO to `5.7.1` (PR #72).